Add files via upload
In this report, we present the results of our unsupervised learning analysis on the Breast Cancer Gene Expression Profiles dataset (METABRIC). The objective of this project was to apply unsupervised learning techniques to cluster data points into 'n' distinct clusters and identify the most important features and their corresponding values contributing significantly to the grouping of data points within these clusters.